Add Intermediate CA Signing Option.
Allow users to generate an intermediate CA certificate in Azure KeyVault
and use this program to sign the intermediate CA with the specified
root TA.

Adds a new option to the Csr configuration: 'IsIntermediateCA'. This
option defaults to 'false'. If set to 'true', the signed certificate
generated from the CSR will have the correct flags set to allow this
certificate to be used as a CA.

Also set default values of 'IsIntermediateCA' and 'IsRootCA' to false,
so we can get away with not specifying them on the command line if we
don't want to.