Add Validity upper bound configuration parameter
Adds a configuration parameter to the EstConfiguration that specifies
the maximum lifetime of a certificate generated by this program. As we
now handle intermediate CA certificates, we need to be able to specify
a configurable validity period, as 365 days is too short for a
Intermediate CA - the default value chosen for this parameter is 5
years.